medio/requirements.txt
Django==1.6.5
MySQL-python==1.2.5
Pillow==2.5.0
South==1.0
django-grappelli==2.5.3
sorl-thumbnail==11.12
tablib==0.10.0
wsgiref==0.1.2